Add command for driving and turning at the same time
Also add a second command for driving to the hopper, taking advantage of
this new ability to drive and turn at the same time; this new command
does not yet work fully. For some reason, the turning does not seem to
kick in until a while after the robot has begun driving forward.