Create separate robot maps for the kit bot and the real robot
RobotMapKitBot contains the constants for the kit bot.
RobotMapReal contains the constants for the real robot.
RobotMap contains the correct constants for whichever robot we are
using; it loads them from either RobotMapKitBot or RobotMapReal
depending on the value of RobotMap.IS_KIT_BOT.

In the future, RobotMap.IS_KIT_BOT should be calculated from some
characteristic of the RoboRIO.  Right now, it is hard-coded. Therefore,
before deploying the code you need to check/change RobotMap.IS_KIT_BOT
appropriately.